+++ apps/client-web/package.json
@@ -18,6 +18,7 @@
     "@ebox/client-api": "workspace:*",
     "@ebox/db": "workspace:*",
     "@ebox/redis-client": "workspace:*",
+    "@ebox/stripe": "workspace:*",
     "@ebox/ui": "workspace:*",
     "@t3-oss/env-nextjs": "^0.9.2",
     "@tanstack/react-query": "^5.29.0",
