+++ apps/client-web/src/app/onboarding/page.tsx
@@ -1,11 +1,25 @@
 import { redirect } from "next/navigation";
-import { hasValidSubscription } from "@/lib/stripe";
 import { api } from "@/trpc/server";
+import { currentUser } from "@clerk/nextjs/server";
+
+import { hasValidSubscription } from "@ebox/stripe";
 
 import PortraitPhotoUpload from "./PortraitPhotoUpload";
 
 export default async function Page() {
-  const userHasValidSubscription = await hasValidSubscription();
+  const user = await currentUser();
+
+  if (!user) {
+    console.error("Request does not have user");
+    return false;
+  }
+
+  const stripeCustomerId = user.privateMetadata.stripeCustomerId as string;
+  if (!stripeCustomerId) {
+    return false;
+  }
+
+  const userHasValidSubscription = await hasValidSubscription(stripeCustomerId);
 
   if (!userHasValidSubscription) {
     redirect("/payment");
