+++ apps/client-web/src/app/success/page.tsx
@@ -1,8 +1,22 @@
 import { redirect } from "next/navigation";
-import { hasValidSubscription } from "@/lib/stripe";
+import { currentUser } from "@clerk/nextjs/server";
+
+import { hasValidSubscription } from "@ebox/stripe";
 
 export default async function SuccessPage() {
-  const userHasValidSubscription = await hasValidSubscription();
+  const user = await currentUser();
+
+  if (!user) {
+    console.error("Request does not have user");
+    return false;
+  }
+
+  const stripeCustomerId = user.privateMetadata.stripeCustomerId as string;
+  if (!stripeCustomerId) {
+    return false;
+  }
+
+  const userHasValidSubscription = await hasValidSubscription(stripeCustomerId);
 
   if (userHasValidSubscription) {
     redirect("/onboarding");
