+++ packages/admin-api/package.json
@@ -21,6 +21,8 @@
   "dependencies": {
     "@clerk/nextjs": "^6.9.3",
     "@ebox/db": "workspace:*",
+    "@ebox/redis-client": "workspace:*",
+    "@ebox/stripe": "workspace:*",
     "@trpc/server": "11.0.0-next.320",
     "jose": "^5.8.0",
     "next-axiom": "^1.4.1",
