+++ packages/admin-api/src/router/orders.ts
@@ -1,6 +1,11 @@
+import { clerkClient } from "@clerk/nextjs/server";
 import { UserType } from "@prisma/client";
 import { TRPCError } from "@trpc/server";
+import Stripe from "stripe";
 import { z } from "zod";
+
+import { kv } from "@ebox/redis-client";
+import { priceIdsToPlan, subscriptionDataSchema } from "@ebox/stripe";
 
 import { createTRPCRouter, protectedAdminProcedure } from "../trpc";
 
@@ -207,6 +212,21 @@ export const ordersRouter = createTRPCRouter({
     )
     .mutation(async ({ ctx, input }) => {
       const { orderId, customerId } = input;
+      const clerk = await clerkClient();
+      const user = await clerk.users.getUser(customerId);
+      if (!user) {
+        throw new TRPCError({
+          code: "UNAUTHORIZED",
+          message: "User not found",
+        });
+      }
+      const stripeCustomerId = user.privateMetadata.stripeCustomerId as string;
+      if (!stripeCustomerId) {
+        throw new TRPCError({
+          code: "UNAUTHORIZED",
+          message: "User does not have a Stripe customer ID",
+        });
+      }
       const order = await ctx.db.order.findUniqueOrThrow({
         where: {
           id: orderId,
@@ -241,6 +261,14 @@ export const ordersRouter = createTRPCRouter({
         });
       }
 
+      if (!order.deliveredDate) {
+        console.error(`Order ID ${input.orderId} was not delivered.`);
+        throw new TRPCError({
+          code: "BAD_REQUEST",
+          message: "This order was not delivered.",
+        });
+      }
+
       await ctx.db.order.update({
         where: {
           id: input.orderId,
@@ -254,5 +282,61 @@ export const ordersRouter = createTRPCRouter({
           },
         },
       });
+
+      const subscriptionDataKv = await kv.get(
+        `stripe:customer:${stripeCustomerId}`,
+      );
+
+      const parsedSubData =
+        subscriptionDataSchema.safeParse(subscriptionDataKv);
+
+      if (!parsedSubData.success) {
+        return false;
+      }
+
+      const subscriptionData = parsedSubData.data;
+      const plan = await priceIdsToPlan(subscriptionData.priceIds);
+      if (!plan) {
+        throw new TRPCError({
+          code: "NOT_FOUND",
+          message: "Unable to determine subscription tier from price IDs",
+        });
+      }
+      const allowedHoldingPeriod =
+        await ctx.db.subscriptionLimit.findUniqueOrThrow({
+          where: {
+            type: plan,
+          },
+          select: {
+            maxPackageHolding: true,
+          },
+        });
+      const maxHoldingDays = allowedHoldingPeriod.maxPackageHolding;
+
+      const numDaysHeld = Math.ceil(
+        (new Date().getTime() - order.deliveredDate.getTime()) /
+          (1000 * 60 * 60 * 24),
+      );
+      const overdueDays = numDaysHeld - maxHoldingDays;
+      if (overdueDays > 0) {
+        // Send Stripe metering event
+        const stripe = new Stripe(process.env.STRIPE_SECRET_KEY!);
+
+        const meterEvent = await stripe.billing.meterEvents.create({
+          event_name: "ferris_credits",
+          payload: {
+            value: overdueDays.toString(),
+            stripe_customer_id: stripeCustomerId,
+          },
+        });
+        await ctx.db.meterEvent.create({
+          data: {
+            eventType: "OVERDUE_PACKAGE_HOLDING",
+            value: overdueDays,
+            customerId: stripeCustomerId,
+            orderId: order.id,
+          },
+        });
+      }
     }),
 });
