+++ packages/client-api/package.json
@@ -22,6 +22,7 @@
     "@clerk/nextjs": "^6.9.3",
     "@ebox/db": "workspace:*",
     "@ebox/redis-client": "workspace:*",
+    "@ebox/stripe": "workspace:*",
     "@trpc/server": "11.0.0-next.320",
     "jose": "^5.8.0",
     "next-axiom": "^1.4.1",
