+++ packages/client-api/src/router/meter.ts
@@ -3,12 +3,9 @@ import { TRPCError } from "@trpc/server";
 import { z } from "zod";
 
 import { kv } from "@ebox/redis-client";
+import { priceIdsToPlan, subscriptionDataSchema } from "@ebox/stripe";
 
 import { createTRPCRouter, protectedCustomerProcedure } from "../trpc";
-import {
-  priceIdsToPlan,
-  subscriptionDataSchema,
-} from "../utils/subscription-utils";
 
 export const meterRouter = createTRPCRouter({
   getCurrentUsage: protectedCustomerProcedure.query(async ({ ctx }) => {
