+++ packages/db/prisma/schema.prisma
@@ -330,6 +330,13 @@ model OrderSharedAccess {
     @@index([grantedById])
 }
 
+model SubscriptionLimit {
+    type              SubscriptionType @unique
+    packageAllowance  Int
+    maxPackageHolding Int
+    locationLimit     Int
+}
+
 enum UserType {
     CUSTOMER
     EMPLOYEE
