+++ packages/stripe/src/index.ts
@@ -1,4 +1,3 @@
-import { currentUser } from "@clerk/nextjs/server";
 import { SubscriptionType } from "@prisma/client";
 import { TRPCError } from "@trpc/server";
 import Stripe from "stripe";
@@ -6,7 +5,6 @@ import z from "zod";
 
 import { kv } from "@ebox/redis-client";
 
-// Must match the type in client-web
 export type SubscriptionData = {
   subscriptionId: string;
   status:
@@ -33,7 +31,6 @@ export type SubscriptionData = {
   };
 };
 
-// Must match the schema in client-web
 export const subscriptionDataSchema = z.object({
   subscriptionId: z.string(),
   status: z.enum([
@@ -165,27 +162,7 @@ export async function priceIdsToPlan(priceIds: string[]) {
 /**
  * Get user's current subscription tier from KV store
  */
-async function getUserSubscriptionTier() {
-  const user = await currentUser();
-
-  if (!user) {
-    throw new TRPCError({
-      code: "UNAUTHORIZED",
-      message: "User not authenticated",
-    });
-  }
-
-  const stripeCustomerId = user.privateMetadata.stripeCustomerId as string;
-
-  if (!stripeCustomerId) {
-    throw new TRPCError({
-      code: "PRECONDITION_FAILED",
-      message:
-        "No Stripe customer ID found. Please set up your subscription first.",
-    });
-  }
-
-  // Get the customer's subscription data from KV store
+async function getUserSubscriptionTier(stripeCustomerId: string) {
   const subscriptionDataKv = await kv.get(
     `stripe:customer:${stripeCustomerId}`,
   );
@@ -229,8 +206,11 @@ async function getUserSubscriptionTier() {
  * - limit: number - The maximum number of favorite locations the user can have
  * - remaining: number - The number of favorite locations the user can add
  */
-export async function canUserAddMoreFavorites(currentFavoriteCount: number) {
-  const tier = await getUserSubscriptionTier();
+export async function canUserAddMoreFavorites(
+  currentFavoriteCount: number,
+  stripeCustomerId: string,
+) {
+  const tier = await getUserSubscriptionTier(stripeCustomerId);
 
   if (!tier) {
     return {
@@ -250,6 +230,7 @@ export async function canUserAddMoreFavorites(currentFavoriteCount: number) {
   };
 }
 
+// TODO: replace with DB query
 export function getLocationLimit(subscriptionTier: SubscriptionType): number {
   return LOCATION_LIMITS[subscriptionTier as keyof typeof LOCATION_LIMITS] || 0;
 }
@@ -304,4 +285,94 @@ export function getScheduledChangeType(
   if (scheduledLevel > currentLevel) return "upgrade";
   if (scheduledLevel < currentLevel) return "downgrade";
   return "none";
+}
+
+/**
+ * Check if user has a valid subscription (status: "active")
+ * Valid even if cancelAtPeriodEnd=true, as long as still in billing period
+ */
+export async function hasValidSubscription(stripeCustomerId: string) {
+  const subscriptionDataKv = await kv.get(
+    `stripe:customer:${stripeCustomerId}`,
+  );
+
+  const parsedSubData = subscriptionDataSchema.safeParse(subscriptionDataKv);
+
+  if (!parsedSubData.success) {
+    return false;
+  }
+
+  const subscriptionData = parsedSubData.data;
+
+  return subscriptionData.status === "active";
+}
+
+// Used in webhook
+export async function syncCustomerData(customerId: string) {
+  const stripe = new Stripe(process.env.STRIPE_SECRET_KEY!);
+
+  const subscriptions = await stripe.subscriptions.list({
+    customer: customerId,
+    limit: 1,
+    status: "all",
+    expand: ["data.default_payment_method"],
+  });
+
+  if (subscriptions.data.length === 0) {
+    const subData = { status: "none" };
+    await kv.set(`stripe:customer:${customerId}`, subData);
+    return subData;
+  }
+
+  const subscription = subscriptions.data[0];
+
+  if (!subscription) {
+    throw new Error(`No subscriptions found`);
+  }
+
+  const subData: SubscriptionData = {
+    subscriptionId: subscription.id,
+    status: subscription.status,
+    priceIds: subscription.items.data.map((item) => item.price.id),
+    currentPeriodEnd: subscription.items.data[0]?.current_period_end!,
+    currentPeriodStart: subscription.items.data[0]?.current_period_start!,
+    cancelAtPeriodEnd: subscription.cancel_at_period_end,
+  };
+
+  if (subscription.schedule && typeof subscription.schedule === "string") {
+    const schedule = await stripe.subscriptionSchedules.retrieve(
+      subscription.schedule,
+    );
+
+    const currentTime = Math.floor(Date.now() / 1000);
+    // Get the next phase that hasn't started yet
+    const upcomingPhase = schedule.phases.find(
+      (phase) => phase.start_date > currentTime,
+    );
+
+    if (upcomingPhase) {
+      console.log("Found upcoming scheduled phase:", {
+        scheduleId: schedule.id,
+        startDate: upcomingPhase.start_date,
+        itemCount: upcomingPhase.items.length,
+      });
+
+      subData.schedule = {
+        scheduleId: schedule.id,
+        startDate: upcomingPhase.start_date,
+        endDate: upcomingPhase.end_date ?? 0, // Default to 0 if no end date
+        items: upcomingPhase.items.map((item) => ({
+          price: typeof item.price === "string" ? item.price : item.price.id,
+        })),
+      };
+    } else {
+      console.log(
+        "No upcoming scheduled phase found for schedule:",
+        schedule.id,
+      );
+    }
+  }
+
+  await kv.set(`stripe:customer:${customerId}`, subData);
+  return subData;
 }
