+++ packages/stripe/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@ebox/tsconfig/internal-package.json",
+  "include": ["src"],
+  "exclude": ["node_modules"]
+}
